# feat(obj): Store properties and methods in a slice rather than a hashmap